[FEATURE] Build optimized Money Tracker API - Classify the test cases
